move variant logic to styles
