+++ frontend/src/metabase/core/components/Link/Link.styled.tsx
@@ -1,13 +1,16 @@
 import styled from "@emotion/styled";
+import { css } from "@emotion/react";
 import { display, space, hover, color } from "styled-system";
 import { Link } from "react-router";
 import { color as metabaseColor } from "metabase/lib/colors";
 import { shouldForwardNonTransientProp } from "metabase/lib/styling/emotion";
 import { focusOutlineStyle } from "metabase/core/style/input";
 
+import type { LinkProps } from "./types";
+
 export const LinkRoot = styled(Link, {
   shouldForwardProp: shouldForwardNonTransientProp,
-})`
+})<LinkProps>`
   ${display};
   ${space};
   ${hover};
@@ -18,17 +21,19 @@ export const LinkRoot = styled(Link, {
   transition: opacity 0.3s linear;
 
   ${focusOutlineStyle("brand")};
+
+  ${props => variants[props.variant ?? "default"] ?? ""}
 `;
 
 export const variants = {
-  default: LinkRoot,
-  brand: styled(LinkRoot)`
+  default: "",
+  brand: css`
     color: ${metabaseColor("brand")};
     &:hover {
       text-decoration: underline;
     }
   `,
-  brandBold: styled(LinkRoot)`
+  brandBold: css`
     color: ${metabaseColor("brand")};
     font-weight: bold;
     &:hover {
