+++ frontend/src/metabase/core/components/Link/types.ts
@@ -1,0 +1,11 @@
+export interface LinkProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
+  to: string;
+  variant?: "default" | "brand" | "brandBold";
+  disabled?: boolean;
+  className?: string;
+  children?: ReactNode;
+  tooltip?: string | TooltipProps;
+  activeClassName?: string;
+  activeStyle?: CSSProperties;
+  onlyActiveOnIndex?: boolean;
+}
